+++ src/views/collection-view/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './format-sidebar-sections'
+export * from './get-slug'
